+++ include/franka/robot_state.h
@@ -327,6 +327,14 @@ struct RobotState {
   std::array<double, 6> O_dP_EE_d{};  // NOLINT(readability-identifier-naming)
 
   /**
+   * \f${^OddP}_O\f$
+   * Linear component of the acceleration of the robot's base, expressed in frame parallel to the
+   * @ref o-frame "base frame", i.e. the base's translational acceleration. If the base is resting
+   * this shows the direction of the gravity vector.
+   */
+  std::array<double, 3> O_ddP_O{};  // NOLINT(readability-identifier-naming)
+
+  /**
    * \f${^OT_{EE}}_{c}\f$
    * Last commanded end effector pose of motion generation in @ref o-frame "base frame".
    * Pose is represented as a 4x4 matrix in column-major format.