+++ src/robot_state.cpp
@@ -75,7 +75,7 @@ std::ostream& operator<<(std::ostream& ostream, const franka::RobotState& robot_
           << ", \"O_F_ext_hat_K\": " << robot_state.O_F_ext_hat_K
           << ", \"K_F_ext_hat_K\": " << robot_state.K_F_ext_hat_K
           << ", \"O_dP_EE_d\": " << robot_state.O_dP_EE_d
-          << ", \"O_T_EE_c\": " << robot_state.O_T_EE_c
+          << ", \"O_ddP_O\": " << robot_state.O_ddP_O << ", \"O_T_EE_c\": " << robot_state.O_T_EE_c
           << ", \"O_dP_EE_c\": " << robot_state.O_dP_EE_c
           << ", \"O_ddP_EE_c\": " << robot_state.O_ddP_EE_c << ", \"theta\": " << robot_state.theta
           << ", \"dtheta\": " << robot_state.dtheta